Fix migrations to explicitly set timestamps fields to be not nullable
Rails 3 behavior set created_at/updated_at to not null if the migration
did not explicitly specify the nullability.
In Rails 4, this behavior changed and these fields now default to null
if not specified. However, Rails 4 raises a deprecation error that Rails 5
will default to not null again (null: false) if you don't specify it.

The safest behavior is to specify null: false for all migrations since
production code running Marty was run as Rails 3 initially, so these
tables have "not null" for the timestamps field already